Revisit Full2VideoRecorder using actions
